Order_Module4

+++ src/main/java/com/bridgelabz/BookStore/dto/CartDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CartDto {
     public String token;
-    public int book_id;
-    public int quantity;
+    public long book_id;
+    public long quantity;
 }
